feat(apollo.config): changed the expected config setting in apollo.config
Apollo.config.splitServices.directory is now exected to provide the full relative directory name and
this package will no longer assume the services are in a "services" directory. Also added a bunch of
tests.

BREAKING CHANGE: Apollo.config.splitServices.directory is expected to be the full relative directory
name and this package will no longer assume that services are in a "services" directory.